requirements.txt
django
django-extensions
pytest
pytest-django
djangorestframework
